double vardiff

+++ components/stratum/include/stratum_api.h
@@ -50,7 +50,7 @@ typedef struct
     uint32_t version;
     uint32_t target;
     uint32_t ntime;
-    uint32_t difficulty;
+    double difficulty;
     bool clean_jobs;
 } mining_notify;
 
@@ -66,7 +66,7 @@ typedef struct
     // mining.notify
     mining_notify *mining_notification;
     // mining.set_difficulty
-    uint32_t new_difficulty;
+    double new_difficulty;
     // mining.set_version_mask
     uint32_t version_mask;
     // result
